refactor drawable
