+++ client/src/pages/Component/Visualizer/css/comments.md
@@ -1,0 +1,3 @@
+The CSS has to go. Either we strict follow the pattern where CSS applies locally or we use global one. 
+
+My bad!
