+++ client/src/pages/Component/Visualizer/parser/treeParser.ts
@@ -1,11 +1,11 @@
 import { GlobalStateStore, useGlobalStore } from '../../../Store/globalStateStore';
 import { UserAnnotation, DataStructureType, isTreeNode } from '../../../Types/annotationType';
 import { Addr, BackendState } from '../../../Types/backendType';
-import { EntityType } from '../../../Types/entity/baseEntity';
-import { EdgeEntity } from '../../../Types/entity/edgeEntity';
-import { NodeEntity, DEFAULT_NODE_SIZE } from '../../../Types/entity/nodeEntity';
+import { EntityType } from '../../../Types/Entity/baseEntity';
+import { EdgeEntity } from '../../../Types/Entity/edgeEntity';
+import { NodeEntity, DEFAULT_NODE_SIZE } from '../../../Types/Entity/nodeEntity';
 import { FrontendTreeGraph, EntityConcrete } from '../../../Types/frontendType';
-import { assertUnreachable } from '../util/util';
+import { assertUnreachable } from '../Util/util';
 import { Parser } from './parser';
 
 type TreeNode = {
